requirements.txt
python-dotenv
requests>=2.25.0
pandas>=1.3.0
pyarrow
python-dateutil
plotly
nltk
Flask==2.2.2
fastapi
uvicorn
httpx
rich
numpy
pydantic
psutil
beautifulsoup4
scikit-learn